Added beginnings of SPI DMA implementation

+++ embedded/azydev_embedded/include/azydev/embedded/dma/common/pool.h
@@ -76,7 +76,7 @@ private:
         Allocation& operator=(const Allocation&);
 
         // IDMANode
-        virtual uint32_t GetBaseAddress_impl() const override final;
+        virtual uintptr_t GetBaseAddress_impl() const override final;
         virtual uint32_t GetNumBeats_impl() const override final;
         virtual bool IsIncrementing_impl() const override final;
         virtual RESULT Reset_impl() override final;
